gRPC: introduce WithIgnoreCancel.
With this enabled, context.Canceled and gprc_codes.Canceled are
considered OK. A separate attribute is set to true if a cancellation
was detected.

To avoid conflating situations where some downstream server component
issued a cancel of a child context, and that is now propagated over
gRPC, we first check if the parent context is canceled.

It could be argued that status.Convert/FromError should use
errors.Unwrap, since the cancelation that is being returned to a
server interceptor was received from a downstream gRPC client. For
the simplicity of this commit, this is left unchanged from before.

Closes #2644.